newsgenie/requirements.txt
streamlit>=1.30.0
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
requests>=2.31.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
